Current

+++ system/scripts/update/run_update_engines_system_software.sh
@@ -3,6 +3,11 @@ touch /opt/engines/run/system/flags/update_engines_running
 
 sudo -n /opt/engines/system/scripts/update/sudo/_update_engines_system_software.sh 
 
+if test -f /opt/engines/run/system/flags/update_pending
+ then
+	rm /opt/engines/run/system/flags/update_pending
+fi
+
 echo "Stopping"
 
 docker stop system 
@@ -42,6 +47,14 @@ docker start system
   		fi
   done 
 touch /opt/engines/run/system/flags/update_engines_run
-rm /opt/engines/run/system/flags/update_engines_running
-rm /opt/engines/run/system/flags/update_pending
+if test -f /opt/engines/run/system/flags/update_engines_running
+ then
+	rm /opt/engines/run/system/flags/update_engines_running
+fi
 
+
+
+if test $1 = '-f'
+ then 
+	docker logs -f system
+ fi
